requirements.txt
transformers==4.39.0
torch==2.2.1+cu118
streamlit==1.32.2
py2neo==2021.2.4
tensorflow>=2.8.0
opencv-python>=4.5.5
scikit_learn==1.4.1.post1
jieba==0.42.1
nltk==3.8.1
pandas==2.2.2
numpy==1.26.4
rouge_chinese==1.0.3
seqeval==1.2.2
tqdm==4.66.2
zhipuai==2.0.1
peft==0.10.0
Pillow>=9.0.0
matplotlib==3.8.3
psutil==5.9.8
pynvml==11.5.0
